move tracepoint
